requirements.txt
# Python 版本要求
python==3.12.4

# PyTorch 及其依赖项
torch==2.0.1+cu118  # CUDA 11.8 版本
torchvision==0.15.2+cu118  # 与 PyTorch 版本匹配的 torchvision 版本
torchaudio==2.0.2+cu118  # 可选的音频支持库

# 或者你可以使用 CPU 版本，如果不需要 CUDA 支持:
# torch==2.0.1+cpu
# torchvision==0.15.2+cpu
# torchaudio==2.0.2+cpu

# 图像处理
Pillow==9.2.0
matplotlib==3.7.2  # 用于绘图
numpy==1.24.2  # 数值计算库

# 数据处理和增强
scikit-learn==1.3.0  # 机器学习工具
pandas==2.1.1  # 数据分析库
